refactor: rename signals to chains
