+++ packages/cerebral-todomvc/package.json
@@ -30,7 +30,7 @@
     "cerebral-module-devtools": "^0.7.0-rc.2",
     "cerebral-module-recorder": "^0.6.0-beta.1",
     "cerebral-module-router": "^0.14.5",
-    "cerebral-view-react": "^0.12.0-beta.4",
+    "cerebral-view-react": "^0.12.0-beta.6",
     "classnames": "^2.2.5",
     "css-loader": "^0.23.1",
     "file-loader": "^0.9.0",
