+++ packages/cerebral-todomvc/src/modules/App/modules/Footer/chains/clearCompleted.js
