+++ packages/cerebral-todomvc/src/modules/App/modules/List/chains/overwriteTodoTitle.js
@@ -1,7 +1,7 @@
 import overwriteTodoTitle from './../actions/overwriteTodoTitle.js'
-import stopEditingTodo from './../actions/stopEditingTodo.js'
+import stopEditingTodo from './stopEditingTodo.js'
 
 export default [
   overwriteTodoTitle,
-  stopEditingTodo
+  ...stopEditingTodo
 ]
