+++ packages/cerebral-todomvc/src/modules/App/modules/List/chains/setTodoNewTitle.js
@@ -1,0 +1,5 @@
+import setTodoNewTitle from './../actions/setTodoNewTitle.js'
+
+export default [
+  setTodoNewTitle
+]
