+++ packages/cerebral-todomvc/src/modules/App/modules/List/chains/stopEditingTodo.js
