+++ packages/cerebral-todomvc/src/modules/App/modules/List/chains/toggleAllChecked.js
