+++ packages/cerebral-todomvc/src/modules/App/modules/List/index.js
@@ -1,10 +1,10 @@
-import newTitleChanged from './signals/newTitleChanged'
-import newTitleSubmitted from './signals/newTitleSubmitted'
-import removeTodoClicked from './signals/removeTodoClicked'
-import todoDoubleClicked from './signals/todoDoubleClicked'
-import toggleAllChanged from './signals/toggleAllChanged'
-import toggleCompletedChanged from './signals/toggleCompletedChanged'
-import newTitleAborted from './signals/newTitleAborted'
+import setTodoNewTitle from './chains/setTodoNewTitle'
+import overwriteTodoTitle from './chains/overwriteTodoTitle'
+import removeTodo from './chains/removeTodo'
+import editTodo from './chains/editTodo'
+import toggleAllChecked from './chains/toggleAllChecked'
+import toggleTodoCompleted from './chains/toggleTodoCompleted'
+import stopEditingTodo from './chains/stopEditingTodo'
 
 export default (options = {}) => {
   return (module) => {
@@ -17,12 +17,16 @@ export default (options = {}) => {
     })
 
     module.addSignals({
-      newTitleChanged,
-      newTitleSubmitted,
-      removeTodoClicked,
-      todoDoubleClicked,
-      toggleAllChanged,
-      toggleCompletedChanged,
-      newTitleAborted})
+      newTitleChanged: {
+        chain: setTodoNewTitle,
+        immediate: true
+      },
+      newTitleSubmitted: overwriteTodoTitle,
+      removeTodoClicked: removeTodo,
+      todoDoubleClicked: editTodo,
+      toggleAllChanged: toggleAllChecked,
+      toggleCompletedChanged: toggleTodoCompleted,
+      newTitleAborted: stopEditingTodo
+    })
   }
 }
