--- packages/cerebral-todomvc/src/modules/App/modules/List/signals/newTitleChanged.js
@@ -1,8 +1,0 @@
-import setTodoNewTitle from './../actions/setTodoNewTitle.js'
-
-export default {
-  chain: [
-    setTodoNewTitle
-  ],
-  immediate: true
-}
