+++ packages/cerebral-todomvc/src/modules/App/modules/NewTodo/chains/changeTitle.js
@@ -1,0 +1,5 @@
+import setTitle from './../actions/setTitle.js'
+
+export default [
+  setTitle
+]
