--- packages/cerebral-todomvc/src/modules/App/modules/NewTodo/signals/titleChanged.js
@@ -1,8 +1,0 @@
-import setTitle from './../actions/setTitle.js'
-
-export default {
-  chain: [
-    setTitle
-  ],
-  immediate: true
-}
