+++ packages/cerebral-todomvc/src/modules/Refs/index.js
@@ -8,8 +8,8 @@ export default (options = {}) => {
 
     module.addServices({
       next (state) {
-        const nextId = state.get([module.name, 'nextRef'])
-        state.set([module.name, 'nextRef'], nextId + 1)
+        const nextId = state.get([module.name, 'nextRef'].join('.'))
+        state.set([module.name, 'nextRef'].join('.'), nextId + 1)
         return nextId
       }
     })
